Move modern_forms coordinator to separate module
