+++ homeassistant/components/modern_forms/__init__.py
@@ -3,36 +3,25 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Coroutine
-from datetime import timedelta
 import logging
 from typing import Any, Concatenate, ParamSpec, TypeVar
 
-from aiomodernforms import (
-    ModernFormsConnectionError,
-    ModernFormsDevice,
-    ModernFormsError,
-)
-from aiomodernforms.models import Device as ModernFormsDeviceState
+from aiomodernforms import ModernFormsConnectionError, ModernFormsError
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.aiohttp_client import async_get_clientsession
 from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.update_coordinator import (
-    CoordinatorEntity,
-    DataUpdateCoordinator,
-    UpdateFailed,
-)
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
+from .coordinator import ModernFormsDataUpdateCoordinator
 
 _ModernFormsDeviceEntityT = TypeVar(
     "_ModernFormsDeviceEntityT", bound="ModernFormsDeviceEntity"
 )
 _P = ParamSpec("_P")
 
-SCAN_INTERVAL = timedelta(seconds=5)
 PLATFORMS = [
     Platform.BINARY_SENSOR,
     Platform.FAN,
@@ -97,37 +86,6 @@ def modernforms_exception_handler(
             _LOGGER.error("Invalid response from API: %s", error)
 
     return handler
-
-
-class ModernFormsDataUpdateCoordinator(DataUpdateCoordinator[ModernFormsDeviceState]):  # pylint: disable=hass-enforce-coordinator-module
-    """Class to manage fetching Modern Forms data from single endpoint."""
-
-    def __init__(
-        self,
-        hass: HomeAssistant,
-        *,
-        host: str,
-    ) -> None:
-        """Initialize global Modern Forms data updater."""
-        self.modern_forms = ModernFormsDevice(
-            host, session=async_get_clientsession(hass)
-        )
-
-        super().__init__(
-            hass,
-            _LOGGER,
-            name=DOMAIN,
-            update_interval=SCAN_INTERVAL,
-        )
-
-    async def _async_update_data(self) -> ModernFormsDevice:
-        """Fetch data from Modern Forms."""
-        try:
-            return await self.modern_forms.update(
-                full_update=not self.last_update_success
-            )
-        except ModernFormsError as error:
-            raise UpdateFailed(f"Invalid response from API: {error}") from error
 
 
 class ModernFormsDeviceEntity(CoordinatorEntity[ModernFormsDataUpdateCoordinator]):
