+++ homeassistant/components/modern_forms/binary_sensor.py
@@ -8,8 +8,9 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.util import dt as dt_util
 
-from . import ModernFormsDataUpdateCoordinator, ModernFormsDeviceEntity
+from . import ModernFormsDeviceEntity
 from .const import CLEAR_TIMER, DOMAIN
+from .coordinator import ModernFormsDataUpdateCoordinator
 
 
 async def async_setup_entry(
