+++ tests/components/modern_forms/test_init.py
@@ -15,7 +15,7 @@ from tests.test_util.aiohttp import AiohttpClientMocker
 
 
 @patch(
-    "homeassistant.components.modern_forms.ModernFormsDevice.update",
+    "homeassistant.components.modern_forms.coordinator.ModernFormsDevice.update",
     side_effect=ModernFormsConnectionError,
 )
 async def test_config_entry_not_ready(
