+++ tests/components/modern_forms/test_light.py
@@ -119,7 +119,9 @@ async def test_light_error(
 
     aioclient_mock.post("http://192.168.1.123:80/mf", text="", status=400)
 
-    with patch("homeassistant.components.modern_forms.ModernFormsDevice.update"):
+    with patch(
+        "homeassistant.components.modern_forms.coordinator.ModernFormsDevice.update"
+    ):
         await hass.services.async_call(
             LIGHT_DOMAIN,
             SERVICE_TURN_OFF,
@@ -139,9 +141,11 @@ async def test_light_connection_error(
     await init_integration(hass, aioclient_mock)
 
     with (
-        patch("homeassistant.components.modern_forms.ModernFormsDevice.update"),
         patch(
-            "homeassistant.components.modern_forms.ModernFormsDevice.light",
+            "homeassistant.components.modern_forms.coordinator.ModernFormsDevice.update"
+        ),
+        patch(
+            "homeassistant.components.modern_forms.coordinator.ModernFormsDevice.light",
             side_effect=ModernFormsConnectionError,
         ),
     ):
